// src/sdk/types.ts
export interface EventGroup {
  id: string
  impl: string
  version: number
  platform: string
  self: {
    user_id: string
    platform: string
    nickname: string
    user_displayname: string
  }
  self_id: number
  time: number
  detail_type: string
  type: string
  sub_type: string
  post_type: string
  message_id: string
  user_id: string
  seq: number
  rand: number
  font: string
  message: {
    type: string
    data: any // You can provide a more specific type if known
  }[]
  raw_message: string
  message_type: string
  sender: {
    user_id: number
    nickname: string
    sub_id: number
    card: string
    sex: string
    age: number
    area: string
    level: number
    role: string
    title: string
  }
  group_id: string
  group_name: string
  block: boolean
  anonymous: null | any // You can provide a more specific type if known
  atme: boolean
  atall: boolean
  group: {
    group_id: number
    group_name: string
    member_count: number
    max_member_count: number
    owner_id: number
    admin_flag: boolean
    last_join_time: number
    last_sent_time: number
    shutup_time_whole: number
    shutup_time_me: number
    create_time: number
    grade: number
    max_admin_count: number
    active_member_count: number
    update_time: number
  }
  member: {
    group_id: number
    user_id: number
    nickname: string
    card: string
    sex: string
    age: number
    join_time: number
    last_sent_time: number
    level: number
    role: string
    title: string
    title_expire_time: number
    shutup_time: number
    update_time: number
    area: string
    rank: string
  }
  cqCode: string
}